Add clean dist, move few funcs to deprecated
