+++ package.json
@@ -5,7 +5,8 @@
     "node": ">=16.*"
   },
   "scripts": {
-    "build": "tsc",
+    "clean": "del-cli dist",
+    "build": "npm run clean && tsc",
     "build:production": "npm run prestart && npm prune --production",
     "watch": "tsc --w",
     "prestart": "npm run build && func extensions install",
@@ -20,6 +21,7 @@
     "@typescript-eslint/eslint-plugin": "^5.42.1",
     "@typescript-eslint/parser": "^5.42.1",
     "azurite": "^3.20.1",
+    "del-cli": "^5.0.0",
     "eslint": "^8.27.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
