+++ tsconfig.json
@@ -10,5 +10,5 @@
     "forceConsistentCasingInFileNames": true,
     "noImplicitAny": false,
   },
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "deprecated"]
 }
